Give each battle its own URL (/srp/battle/:id)
Start date in battle row now links to detail. Doesn't include any
extra information, but allows people to link to specific battles.
